1) Preserve None values saved in the SQL table 2) Fix combination of temporary dataframe and SQL table
1) Turns out that if a None value was saved in the SQL table, then
   combine_first didn't always replace the corresponding NaN value in
the temporary dataframe by this None, since None is another type of null
value.  Now, the temporary dataframe is filled with None rather than
NaN, so if a None is saved in the SQL table, then it is preserved.

2) There was a bug: we were replacing the values in the SQL table with
   the values in the temporary dataframe, rather than doing it the other
way around. The correct way is:
- Preserve the non-null values in the temporary dataframe
- Substitute the null values in the temporary dataframe by the
corresponding values in the SQL table.